Remove incorrect usage

+++ nav2_smac_planner/include/nav2_smac_planner/collision_checker.hpp
@@ -89,7 +89,7 @@ public:
    * @return boolean if in collision or not.
    */
   bool inCollision(
-    const uint64_t & i,
+    const unsigned int & i,
     const bool & traverse_unknown);
 
   /**
